fix(pdf): accept double tax rates in receipt models

+++ lib/src/cezeri_pdf_helper/models/receipt_product.dart
@@ -4,7 +4,7 @@ class CZRReceiptProduct {
   final double unitPriceNet;
   final double unitPriceGross;
   final int quantity;
-  final int taxRate;
+  final double taxRate;
 
   const CZRReceiptProduct({
     required this.articleNumber,
